Move createMetadata function from client to script
